pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-autobidder-de"
version = "0.1.0"
description = "Tesla Autobidder-inspired BESS optimization engine for the German power market."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAutobidder-DE Contributors" }
]
dependencies = [
  "pandas>=2.2.0",
  "pyarrow>=16.0.0",
  "numpy>=1.26.0",
  "scipy>=1.12.0",
  "pulp>=2.8.0",
  "streamlit>=1.34.0",
  "plotly>=5.20.0",
  "pydantic>=2.6.0",
  "entsoe-py>=0.6.18",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
